% Auto-generated: do not edit by hand
\name{''DraggableDiv}

\alias{''DraggableDiv}

\title{DraggableDiv component}

\description{

}

\usage{
''DraggableDiv(children=NULL, id=NULL, style=NULL)
}

\arguments{
\item{children}{The children of this component}

\item{id}{The ID of this component, used to identify dash components in callbacks}

\item{style}{The style of this component}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

