% Auto-generated: do not edit by hand
\name{''GridItem}

\alias{''GridItem}

\title{GridItem component}

\description{

}

\usage{
''GridItem(children=NULL, id=NULL, bounded=NULL, draggable=NULL,
height=NULL, max_height=NULL, max_width=NULL,
min_height=NULL, min_width=NULL, resizable=NULL,
static=NULL, style=NULL, width=NULL, x=NULL, y=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of the grid}

\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{bounded}{Logical. Whether the grid item is bounded within the grid container}

\item{draggable}{Logical. Whether the grid item is draggable}

\item{height}{Numeric. The height of the grid item}

\item{max_height}{Numeric. The maximum height of the grid item}

\item{max_width}{Numeric. The maximum width of the grid item}

\item{min_height}{Numeric. The minimum height of the grid item}

\item{min_width}{Numeric. The minimum width of the grid item}

\item{resizable}{Logical. Whether the grid item is resizable}

\item{static}{Logical. Whether the grid item is static (non-draggable and non-resizable)}

\item{style}{Named list. The style of the grid item}

\item{width}{Numeric. The width of the grid item}

\item{x}{Numeric. The x position of the grid item}

\item{y}{Numeric. The y position of the grid item}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

