% Auto-generated: do not edit by hand
\name{''Grid}

\alias{''Grid}

\title{Grid component}

\description{

}

\usage{
''Grid(children=NULL, id=NULL, cols=NULL, isDraggable=NULL,
isResizable=NULL, layout=NULL, onLayoutChange=NULL,
rowHeight=NULL)
}

\arguments{
\item{children}{The children of the grid}

\item{id}{The ID used to identify this component in Dash callbacks}

\item{cols}{The number of columns in the grid}

\item{isDraggable}{Whether the grid items are draggable}

\item{isResizable}{Whether the grid items are resizable}

\item{layout}{The layout of the grid
This is an array of objects, see the demo for more complete usage}

\item{onLayoutChange}{The callback that is fired when the layout changes}

\item{rowHeight}{The row height of the grid}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

