% Auto-generated: do not edit by hand
\name{''Grid}

\alias{''Grid}

\title{Grid component}

\description{

}

\usage{
''Grid(children=NULL, id=NULL, allowOverlap=NULL, autoSize=NULL,
cols=NULL, compactType=NULL, containerPadding=NULL,
isDraggable=NULL, isResizable=NULL, layout=NULL,
margin=NULL, preventCollision=NULL, resizeHandles=NULL,
rowHeight=NULL, width=NULL)
}

\arguments{
\item{children}{The children of the grid}

\item{id}{The ID used to identify this component in Dash callbacks}

\item{allowOverlap}{allow overlapping in the grid}

\item{autoSize}{If true, container size swells to fit contents}

\item{cols}{The number of columns in the grid}

\item{compactType}{Compaction type}

\item{containerPadding}{Padding inside the container [x, y] in px}

\item{isDraggable}{Whether the grid items are draggable}

\item{isResizable}{Whether the grid items are resizable}

\item{layout}{The layout of the grid, Readonly.}

\item{margin}{Margin between items [x, y] in px}

\item{preventCollision}{If true, grid items wont change position when being dragged over}

\item{resizeHandles}{Which resize handles to display}

\item{rowHeight}{The row height of the grid}

\item{width}{width of the grid}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

